src/main/java/com/kucukcinar/models/Pizza.java
package com.kucukcinar.models;

import java.util.ArrayList;
import java.util.List;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.index.IndexDirection;
import org.springframework.data.mongodb.core.index.Indexed;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "Pizzas")
public class Pizza {

    @Id
    private String id;
    private String name;
    @Indexed(direction = IndexDirection.ASCENDING)
    private int price;
    private List<String> ingredients;
    private String image;

    public Pizza() {
        this.ingredients = new ArrayList<>();
    }

    public Pizza(String name, int price, List<String> ingredients, String image) {
        super();
        this.name = name;
        this.price = price;
        this.ingredients = ingredients;
        this.image = image;
    }

    public String getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getPrice() {
        return price;
    }

    public void setPrice(int price) {
        this.price = price;
    }

    public List<String> getIngredients() {
        return ingredients;
    }

    public void setIngredients(List<String> ingredients) {
        this.ingredients = ingredients;
    }

    public String getImage() {
        return image;
    }

    public void setImage(String image) {
        this.image = image;
    }

    @Override
    public String toString() {
        return "name=" + name + ", price=" + price + ", ingredients=" + ingredients;
    }
}
